+++ sat/apiclient/gateway.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2019-2021 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2019-2022 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -24,6 +24,7 @@
 """
 Client for querying the API gateway.
 """
+from functools import wraps
 import logging
 import requests
 from urllib.parse import urlunparse
@@ -31,6 +32,28 @@ from urllib.parse import urlunparse
 from sat.config import get_config_value
 
 LOGGER = logging.getLogger(__name__)
+
+
+def handle_api_errors(fn):
+    """Decorator to handle common errors with API access.
+
+    This decorator handles APIErrors, ValueErrors, and KeyErrors, raising all
+    as APIErrors with appropriate prefixes determined from the name of the
+    decorated function.
+    """
+    err_prefix = f'Failed to {fn.__name__.replace("_", " ")}'
+
+    @wraps(fn)
+    def inner(*args, **kwargs):
+        try:
+            return fn(*args, **kwargs)
+        except APIError as err:
+            raise APIError(f'{err_prefix}: {err}') from err
+        except ValueError as err:
+            raise APIError(f'{err_prefix} due to bad JSON in response: {err}') from err
+        except KeyError as err:
+            raise APIError(f'{err_prefix} due to missing {err} key in response.') from err
+    return inner
 
 
 class APIError(Exception):