+++ tests/test_hms_discovery.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2020-2021 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2020-2022 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -33,7 +33,12 @@ from dateutil.tz import tzutc
 from kubernetes.config import ConfigException
 from kubernetes.client.rest import ApiException
 
-from sat.cli.bootsys.discovery import HMSDiscoveryCronJob, HMSDiscoveryError, HMSDiscoveryScheduledWaiter
+from sat.hms_discovery import (
+    HMSDiscoveryCronJob,
+    HMSDiscoveryError,
+    HMSDiscoveryScheduledWaiter,
+    HMSDiscoverySuspendedWaiter,
+)
 
 
 class TestHMSDiscoveryCronJob(unittest.TestCase):
@@ -41,8 +46,8 @@ class TestHMSDiscoveryCronJob(unittest.TestCase):
 
     def setUp(self):
         """Set up some mocks."""
-        self.mock_load_config = patch('sat.cli.bootsys.discovery.load_kube_config').start()
-        self.mock_batch_api_cls = patch('sat.cli.bootsys.discovery.BatchV1beta1Api').start()
+        self.mock_load_config = patch('sat.hms_discovery.load_kube_config').start()
+        self.mock_batch_api_cls = patch('sat.hms_discovery.BatchV1beta1Api').start()
         self.mock_batch_api = self.mock_batch_api_cls.return_value
         self.hdcj = HMSDiscoveryCronJob()
 
@@ -107,7 +112,7 @@ class TestHMSDiscoveryCronJob(unittest.TestCase):
             self.hdcj.get_suspend_status()
         )
 
-    @patch('sat.cli.bootsys.discovery.HMSDiscoveryCronJob.get_suspend_status')
+    @patch('sat.hms_discovery.HMSDiscoveryCronJob.get_suspend_status')
     def test_suspend_already_suspended(self, mock_get_suspend_status):
         """Test suspending when already suspended."""
         mock_get_suspend_status.return_value = True
@@ -115,7 +120,7 @@ class TestHMSDiscoveryCronJob(unittest.TestCase):
 
         self.mock_batch_api.patch_namespaced_cron_job.assert_not_called()
 
-    @patch('sat.cli.bootsys.discovery.HMSDiscoveryCronJob.get_suspend_status')
+    @patch('sat.hms_discovery.HMSDiscoveryCronJob.get_suspend_status')
     def test_suspend_already_not_suspended(self, mock_get_suspend_status):
         """Test setting suspend status to False when already not suspended."""
         mock_get_suspend_status.return_value = False
@@ -123,7 +128,7 @@ class TestHMSDiscoveryCronJob(unittest.TestCase):
 
         self.mock_batch_api.patch_namespaced_cron_job.assert_not_called()
 
-    @patch('sat.cli.bootsys.discovery.HMSDiscoveryCronJob.get_suspend_status')
+    @patch('sat.hms_discovery.HMSDiscoveryCronJob.get_suspend_status')
     def test_suspend_when_not_suspended(self, mock_get_suspend_status):
         """Test suspending when not currently suspended."""
         mock_get_suspend_status.return_value = False
@@ -135,7 +140,7 @@ class TestHMSDiscoveryCronJob(unittest.TestCase):
             {'spec': {'suspend': True}}
         )
 
-    @patch('sat.cli.bootsys.discovery.HMSDiscoveryCronJob.get_suspend_status')
+    @patch('sat.hms_discovery.HMSDiscoveryCronJob.get_suspend_status')
     def test_resume_when_suspended(self, mock_get_suspend_status):
         """Test resuming when currently suspended."""
         mock_get_suspend_status.return_value = True
@@ -147,7 +152,7 @@ class TestHMSDiscoveryCronJob(unittest.TestCase):
             {'spec': {'suspend': False}}
         )
 
-    @patch('sat.cli.bootsys.discovery.HMSDiscoveryCronJob.get_suspend_status')
+    @patch('sat.hms_discovery.HMSDiscoveryCronJob.get_suspend_status')
     def test_set_suspend_api_exception(self, mock_get_suspend_status):
         """Test suspending when already suspended."""
         mock_get_suspend_status.return_value = False
@@ -169,7 +174,7 @@ class TestHMSDiscoveryCronJob(unittest.TestCase):
         self.mock_batch_api.read_namespaced_cron_job.return_value.spec = Mock(schedule='*/3 * * * *')
         expected = datetime(2020, 12, 31, 10, 3, 0)
 
-        with patch('sat.cli.bootsys.discovery.datetime', MockDateTime):
+        with patch('sat.hms_discovery.datetime', MockDateTime):
             # Call it twice to ensure it returns the same thing the second time
             # rather than the next expected schedule time.
             next_time = self.hdcj.get_latest_next_schedule_time()
@@ -178,13 +183,26 @@ class TestHMSDiscoveryCronJob(unittest.TestCase):
         self.assertEqual(expected, next_time)
         self.assertEqual(expected, same_next_time)
 
+    def test_is_not_active_with_no_jobs(self):
+        """Test that the job is considered not active if there are not running jobs"""
+        self.mock_batch_api.read_namespaced_cron_job.return_value.status.active = []
+        self.assertFalse(self.hdcj.is_active())
+
+    def test_is_active_when_jobs_active(self):
+        """Test that the job is considered active if there are running jobs"""
+        self.mock_batch_api.read_namespaced_cron_job.return_value.status.active = [
+            Mock(name=f'discovery-job-{n}')
+            for n in range(5)
+        ]
+        self.assertTrue(self.hdcj.is_active())
+
 
 class TestHMSDiscoveryScheduledWaiter(unittest.TestCase):
     """Test the HMSDiscoveryScheduledWaiter class."""
 
     def setUp(self):
         """Set up mocks."""
-        self.mock_datetime = patch('sat.cli.bootsys.discovery.datetime').start()
+        self.mock_datetime = patch('sat.hms_discovery.datetime').start()
 
         self.poll_interval = 5
         self.grace_period = 30
@@ -192,7 +210,7 @@ class TestHMSDiscoveryScheduledWaiter(unittest.TestCase):
         self.mock_datetime.now.return_value = self.start_time
 
         self.next_sched_time = datetime(2020, 10, 31, 8, 3, 0, tzinfo=tzutc())
-        self.mock_hd_cron_job = patch('sat.cli.bootsys.discovery.HMSDiscoveryCronJob').start().return_value
+        self.mock_hd_cron_job = patch('sat.hms_discovery.HMSDiscoveryCronJob').start().return_value
         self.mock_hd_cron_job.get_latest_next_schedule_time.return_value = self.next_sched_time
 
         self.hd_waiter = HMSDiscoveryScheduledWaiter(self.poll_interval, self.grace_period)
@@ -244,6 +262,30 @@ class TestHMSDiscoveryScheduledWaiter(unittest.TestCase):
         self.assertEqual(logging.getLevelName(logging.WARNING), logs.records[0].levelname)
         self.assertEqual(f'Failed to get last schedule time: {msg}',
                          logs.records[0].message)
+
+
+class TestHMSDiscoverySuspendedWaiter(unittest.TestCase):
+    """Tests for the HMSDiscoverySuspendedWaiter class"""
+    def setUp(self):
+        self.mock_hd_cron_job = patch('sat.hms_discovery.HMSDiscoveryCronJob').start().return_value
+
+    def test_cron_job_is_suspended(self):
+        """Test that the cron job is suspended and has no running jobs"""
+        self.mock_hd_cron_job.get_suspend_status.return_value = True
+        self.mock_hd_cron_job.is_active.return_value = False
+        self.assertTrue(HMSDiscoverySuspendedWaiter(10).has_completed())
+
+    def test_cron_job_suspended_with_running_jobs(self):
+        """Test that the waiter continues waiting when the cron job is suspended and but has running jobs"""
+        self.mock_hd_cron_job.get_suspend_status.return_value = True
+        self.mock_hd_cron_job.is_active.return_value = True
+        self.assertFalse(HMSDiscoverySuspendedWaiter(10).has_completed())
+
+    def test_cron_job_not_suspended(self):
+        """Test that the waiter continues waiting when the cron job is not suspended"""
+        for is_active in [False, True]:
+            with self.subTest(is_active=is_active):
+                self.assertFalse(HMSDiscoverySuspendedWaiter(10).has_completed())
 
 
 if __name__ == '__main__':