+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added information regarding the `--bos-version` command line argument to man
   pages for relevant subcommands.
+- Added a ``sat swap blade`` subcommand which partially automates the procedure
+  for swapping compute and UAN blades.
 
 ### Fixed
 - Fixed unit tests that failed when run in PyCharm.